Combined 3 tests here
first test -> to have a no posts found: in index blade I am checking if there is a post or not
second test -> if not a published post get 404: here I added a table in the migrations in order to check if a post is published or not. if its timestamp its published if null its not. then in factory i added the space for the unpublished posts to exist. Then in PostController the logic that checks if there is a post is it publist or not.
third test-> If there is an image to publish: Here in postController we added to the existing logic the filter for the image. It will only show posts that contain image